+++ plugins/forms/lang/tt.js
@@ -1,0 +1,68 @@
+﻿/*
+Copyright (c) 2003-2014, CKSource - Frederico Knabben. All rights reserved.
+For licensing, see LICENSE.md or http://ckeditor.com/license
+*/
+CKEDITOR.plugins.setLang( 'forms', 'tt', {
+	button: {
+		title: 'Төймә үзлекләре',
+		text: 'Text (Value)', // MISSING
+		type: 'Төр',
+		typeBtn: 'Төймә',
+		typeSbm: 'Submit', // MISSING
+		typeRst: 'Reset' // MISSING
+	},
+	checkboxAndRadio: {
+		checkboxTitle: 'Checkbox Properties', // MISSING
+		radioTitle: 'Radio Button Properties', // MISSING
+		value: 'Value', // MISSING
+		selected: 'Сайланган'
+	},
+	form: {
+		title: 'Form Properties', // MISSING
+		menu: 'Form Properties', // MISSING
+		action: 'Action', // MISSING
+		method: 'Method', // MISSING
+		encoding: 'Кодировка'
+	},
+	hidden: {
+		title: 'Hidden Field Properties', // MISSING
+		name: 'Исем',
+		value: 'Күләм'
+	},
+	select: {
+		title: 'Selection Field Properties', // MISSING
+		selectInfo: 'Select Info', // MISSING
+		opAvail: 'Available Options', // MISSING
+		value: 'Күләм',
+		size: 'Зурлык',
+		lines: 'lines', // MISSING
+		chkMulti: 'Allow multiple selections', // MISSING
+		opText: 'Текст',
+		opValue: 'Күләм',
+		btnAdd: 'Кушу',
+		btnModify: 'Үзгәртү',
+		btnUp: 'Up', // MISSING
+		btnDown: 'Down', // MISSING
+		btnSetValue: 'Set as selected value', // MISSING
+		btnDelete: 'Бетерү'
+	},
+	textarea: {
+		title: 'Textarea Properties', // MISSING
+		cols: 'Баганалар',
+		rows: 'Юллар'
+	},
+	textfield: {
+		title: 'Text Field Properties', // MISSING
+		name: 'Исем',
+		value: 'Күләм',
+		charWidth: 'Character Width', // MISSING
+		maxChars: 'Maximum Characters', // MISSING
+		type: 'Төр',
+		typeText: 'Текст',
+		typePass: 'Сер сүз',
+		typeEmail: 'Эл. почта',
+		typeSearch: 'Эзләү',
+		typeTel: 'Телефон номеры',
+		typeUrl: 'Сылталама'
+	}
+} );