+++ plugins/liststyle/lang/tt.js
@@ -1,0 +1,25 @@
+﻿/*
+Copyright (c) 2003-2014, CKSource - Frederico Knabben. All rights reserved.
+For licensing, see LICENSE.md or http://ckeditor.com/license
+*/
+CKEDITOR.plugins.setLang( 'liststyle', 'tt', {
+	armenian: 'Armenian numbering', // MISSING
+	bulletedTitle: 'Маркерлы тезмә үзлекләре',
+	circle: 'Түгәрәк',
+	decimal: 'Унарлы (1, 2, 3, ...)',
+	decimalLeadingZero: 'Ноль белән башланган унарлы (01, 02, 03, ...)',
+	disc: 'Диск',
+	georgian: 'Georgian numbering (an, ban, gan, etc.)', // MISSING
+	lowerAlpha: 'Lower Alpha (a, b, c, d, e, etc.)', // MISSING
+	lowerGreek: 'Lower Greek (alpha, beta, gamma, etc.)', // MISSING
+	lowerRoman: 'Lower Roman (i, ii, iii, iv, v, etc.)', // MISSING
+	none: 'Һичбер',
+	notset: '<билгеләнмәгән>',
+	numberedTitle: 'Numbered List Properties', // MISSING
+	square: 'Шакмак',
+	start: 'Башлау',
+	type: 'Төр',
+	upperAlpha: 'Upper Alpha (A, B, C, D, E, etc.)', // MISSING
+	upperRoman: 'Upper Roman (I, II, III, IV, V, etc.)', // MISSING
+	validateStartNumber: 'List start number must be a whole number.' // MISSING
+} );