+++ plugins/smiley/lang/tt.js
@@ -1,0 +1,9 @@
+﻿/*
+Copyright (c) 2003-2014, CKSource - Frederico Knabben. All rights reserved.
+For licensing, see LICENSE.md or http://ckeditor.com/license
+*/
+CKEDITOR.plugins.setLang( 'smiley', 'tt', {
+	options: 'Смайл көйләүләре',
+	title: 'Смайл өстәү',
+	toolbar: 'Смайл'
+} );